Isolate the New chat focus lease
